Refactor style suggestions
Follow some minor style suggestions from go fmt and refactor some
functions to utils.go
